feat: added keyboard support and fixed logic for equals button
I created an event listener for the window which detects when the user presses a key
on their keyboard that matches the available functions. Then it will
execute the corresponding logic. I also had to change the equals button
logic because it wasn't storing the result value in the array.